refactor: update package structure and improve metric collection in tests
