+++ observability/metric/integration_test.go
@@ -1,6 +1,6 @@
 //go:build integration
 
-package observability
+package metric
 
 import (
 	"context"
