+++ observability/metric/meter.go
@@ -1,4 +1,4 @@
-package observability
+package metric
 
 import (
 	"context"
